Introduce new home components (weekly progress, goal progress).
Change material colors

Better support for light/dark themes

+++ analytics/data/src/main/java/com/skymonkey/analytics/data/RoomAnalyticsRepository.kt
@@ -20,7 +20,7 @@ class RoomAnalyticsRepository(
             val avgPacePerRun = async { analyticsDao.getAveragePacePerRun() }
 
             AnalyticsValues(
-                totalDistanceRun = totalDistance.await(),
+                totalDistanceRun = totalDistance.await().toInt(),
                 totalTimeRun = totalTimeMillis.await().milliseconds,
                 fastestEverRun = maxRunSpeed.await(),
                 avgDistanceRun = avgDistancePerRun.await(),
